[tests] Create `setup.jl` instead of passing `init_code` to `runtests`
