+++ test/all_sky_radiative_transfer.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Breeze
 using Dates
 using GPUArraysCore: @allowscalar
