+++ test/anelastic_pressure_solver_analytic.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Test
 using Breeze
 using Oceananigans
