+++ test/atmosphere_model_callbacks.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Breeze
 using Oceananigans
 using Oceananigans: UpdateStateCallsite, TendencyCallsite, TimeStepCallsite
