+++ test/boundary_conditions.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Test
 using Breeze
 using Oceananigans.BoundaryConditions: BoundaryCondition
