+++ test/cloud_microphysics_1M.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Breeze
 using Breeze.AtmosphereModels: microphysical_velocities
 using CloudMicrophysics
