+++ test/dcmip2016_kessler.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Breeze
 using Test
 using Oceananigans
