+++ test/doctests.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 # Loading `Breeze` into `Main` is necessary to work around
 # <https://github.com/JuliaTesting/ParallelTestRunner.jl/issues/68>.
 @eval Main using Breeze
