+++ test/dynamics.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Breeze
 using Breeze.Microphysics
 using Oceananigans
