+++ test/instantaneous_precipitation.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Breeze
 using Oceananigans
 using Test
