+++ test/parcel_dynamics.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 #####
 ##### Unit tests for ParcelDynamics module
 #####
