+++ test/polynomial_bulk_coefficients.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Test
 using Breeze
 using Breeze.BoundaryConditions
