+++ test/radiation_scheduling.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Breeze
 using Dates: DateTime, Second
 using GPUArraysCore: @allowscalar
