+++ test/reactant/lat_lon_compilation.jl
@@ -1,3 +1,5 @@
+include(joinpath(dirname(@__DIR__), "setup.jl"))
+
 #####
 ##### Reactant compilation tests — LatitudeLongitudeGrid (Periodic, Bounded, Bounded)
 #####
