+++ test/set_atmosphere_model.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Breeze
 using GPUArraysCore: @allowscalar
 using Oceananigans
