+++ test/substepper_structural.jl
@@ -1,3 +1,5 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 #####
 ##### Structural correctness tests for the acoustic substepper.
 #####
