+++ test/terrain_following_split_explicit.jl
@@ -1,12 +1,7 @@
+include(joinpath(@__DIR__, "setup.jl"))
+
 using Breeze
 using Breeze.AtmosphereModels: transport_velocities
-using Oceananigans.Architectures: CPU
-
-# Run under `default_arch` when the test runner provides it (which routes to
-# GPU when CUDA is functional, matching project convention); otherwise fall
-# back to CPU() so this file can also be included directly with
-# `julia --project=. test/<this file>.jl` during development.
-@isdefined(default_arch) || (default_arch = CPU())
 
 using CUDA: @allowscalar
